refactor(docfx): split `IncludeInPublicDocuments()`
