+++ docfx/public_docs_test.cc
@@ -1,0 +1,66 @@
+// Copyright 2023 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include "docfx/public_docs.h"
+#include <gmock/gmock.h>
+
+namespace docfx {
+namespace {
+
+TEST(PublicDocs, IncludeNoConfig) {
+  auto constexpr kXml = R"xml(<?xml version="1.0" standalone="yes"?>
+    <doxygen version="1.9.1" xml:lang="en-US">
+      <sectiondef id="id-1" kind="private-attrib"></sectiondef>
+      <sectiondef id="id-2" kind="private-func"></sectiondef>
+      <sectiondef id="id-3" kind="public-attrib"></sectiondef>
+      <sectiondef id="id-4" kind="public-func"></sectiondef>
+      <compounddef id="id-5" kind="file" language="C++"></compounddef>
+      <compounddef id="id-6" kind="dir"></compounddef>
+      <compounddef id="namespacestd" kind="namespace"></compounddef>
+      <compounddef id="namespacestd_1_1chrono" kind="namespace"></compounddef>
+      <compounddef id="classstd_1_1array"></compounddef>
+      <compounddef id="classgoogle_1_1cloud_1_1Options" prot="public">google::cloud::Options</compounddef>
+      <compounddef id="classgoogle_1_1cloud_1_1Options_1_1DataHolder" prot="private"></compounddef>
+    </doxygen>)xml";
+  pugi::xml_document doc;
+  doc.load_string(kXml);
+
+  struct TestCase {
+    std::string id;
+    bool expected;
+  } const cases[] = {
+      {"id-1", false},
+      {"id-2", false},
+      {"id-3", true},
+      {"id-4", true},
+      {"id-5", false},
+      {"id-6", false},
+      {"namespacestd", false},
+      {"namespacestd_1_1chrono", false},
+      {"classstd_1_1array", false},
+      {"classgoogle_1_1cloud_1_1Options", true},
+      {"classgoogle_1_1cloud_1_1Options_1_1DataHolder", false},
+  };
+
+  for (auto const& test : cases) {
+    SCOPED_TRACE("Running with id=" + test.id);
+    auto const filter = "//*[@id='" + test.id + "']";
+    auto selected = doc.select_node(filter.c_str());
+    ASSERT_TRUE(selected);
+    EXPECT_EQ(test.expected, IncludeInPublicDocuments(selected.node()));
+  }
+}
+
+}  // namespace
+}  // namespace docfx
